requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
colorama==0.4.4
Faker==8.9.1
pytest==7.1.2
pytest-forked==1.6.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytest-check==2.3.1
requests==2.28.1
text-unidecode==1.3
virtualenv==20.14.1
selenium==4.19.0
flask
subprocess32